hw1/homew1/requirements.txt
baidu-aip==2.2.18.0
PyAudio==0.2.11
PyQt5==5.15.2
pywin32==300
qt5-tools==5.15.2.3.0.2